+++ plugins/postcss-nesting/src/lib/is-type-of-rule.ts
@@ -8,10 +8,6 @@ export function isNestRule(node?: Node): node is AtRule {
 	return node && isAtRule(node) && node.name === 'nest';
 }
 
-export function isMixinRule(node?: Node): node is AtRule {
-	return node && isAtRule(node) && node.name.toLowerCase() === 'mixin';
-}
-
 export function isRule(node?: Node): node is Rule {
 	return node && node.type === 'rule';
 }