+++ plugins/postcss-nesting/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changes to PostCSS Nesting
 
-### Unreleased (patch)
+### 12.0.4
+
+_February 26, 2024_
 
 - Fix order of final CSS with complex usage of both nesting and mixins, by @pciarach